requirements.txt
streamlit
openai
python-dotenv
langchain-openai
langchain-core
langgraph
pydantic